backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "alembic>=1.18.4",
    "email-validator>=2.3.0",
    "fastapi>=0.134.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.47",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "uvicorn>=0.41.0",
]
